fix(grammar): Correct Flaw in Expression Statement Grammar in left hand primitive expression.
Previously, the grammar allowed a left primitive expression and equal sign independently of
each other. This caused problems with function calls, where the primitive atom identifier
expression was stripped away from the rest of the call arguments, giving a strange output.
Here, we link the optional left node to require both the primitvie expression and equal sign
to prevent this odd parsing error.